Add stock service base class and models
